Check for cached consumption first
If a Miele device is temporarily disconnected from the Miele cloud while
running, the Miele API returns a "not connected" state for the device. If the
"not running" check is done first, it causes the energy/water consumption to be
reset to 0. This in turn means that the statistics generated by Home Assistant
will be messed up, it will look like more energy/water was consumed than in
reality.