Moving from dev to stable (#110)
* Check for cached consumption first

If a Miele device is temporarily disconnected from the Miele cloud while
running, the Miele API returns a "not connected" state for the device. If the
"not running" check is done first, it causes the energy/water consumption to be
reset to 0. This in turn means that the statistics generated by Home Assistant
will be messed up, it will look like more energy/water was consumed than in
reality.

* Fix missing ecoFeedback object

* fix the coroutine error (#108)

* Bump version

* Switching back to manual versioning

* fix the coroutine error

* warn if unable to delete the cached token
check if authorized before creating a new session

Co-authored-by: sebltm
Co-authored-by: kloknibor <kloknibor@users.noreply.github.com>

Co-authored-by: Sebastian Lövdahl <slovdahl@hibox.fi>
Co-authored-by: mundschenk_at <github@mundschenk.at>
Co-authored-by: Sébastien Michel <26392528+sebltm@users.noreply.github.com>
Co-authored-by: kloknibor <kloknibor@users.noreply.github.com>